Etape 28 : <AffichageReservation_Load> dana le 3e formulaire, le bon label et l'image apparait. De plus, la liste des réservations s'affiche bien

+++ SarahMerzoukTP02/Reservation.cs
@@ -80,7 +80,7 @@ namespace SarahMerzoukTP02
             return _nbEnfants;
         }
 
-        public float getCoutTotal()
+        public decimal getCoutTotal()
         {
             return _coutTotal;
         }
